Status: When substituting into a UTP pack an assert occurs. At least when the pack is the template parameter of a class template specialization.
Added a Pattern member to UniversalTemplateParmDecl which is also a UniversalTemplateParmDecl but with the Pattern member nulled.

Moved the Kind member out of all the union substructs in TemplateArgument. This simplifies natvis and the code in general. It was an unsigned in all union members anyway so no size difference.

Moved the TypeArg member out of the union too as it is used by many of the union members which simplified some code (and natvis). No size difference as the largest union members did have the TypeArg anyway.

Removed the TV struct as it obscured the pointer type which made debugging harder.

Improved (?) natvis. Now you at least se a little more in VS2022.

The change made in llvm.natvis indicates that the natvis files have not been maintained for a long time.